colocate and rename include statements (#145)
