+++ src/lib/hooks/useUpload.ts
@@ -1,9 +1,20 @@
 import { useState } from 'react'
 
-import { Asset, fileToAsset } from '../utils'
-
 const sizeLimit = 1028 * 1028 //1 mb
 const fileLimit = 10
+
+export type Asset = ReturnType<typeof fileToAsset>
+const fileToAsset = (rawFile: File) => {
+  return {
+    size: rawFile.size / 1000,
+    createdAt: new Date(rawFile.lastModified).toISOString(),
+    name: rawFile.name,
+    url: URL.createObjectURL(rawFile),
+    ext: rawFile.name.split('.').pop(),
+    mime: rawFile.type,
+    rawFile,
+  }
+}
 
 const useUpload = <T>(
   files: T[],
